create pytest_indexd.db
