+++ .secrets.baseline
@@ -163,13 +163,13 @@
         "is_secret": false
       }
     ],
-    "pytest_indexd/__init__.py": [
+    "pytest_indexd/db.py": [
       {
         "type": "Secret Keyword",
-        "filename": "pytest_indexd/__init__.py",
+        "filename": "pytest_indexd/db.py",
         "hashed_secret": "d033e22ae348aeb5660fc2140aec35850c4da997",
         "is_verified": false,
-        "line_number": 80,
+        "line_number": 78,
         "is_secret": false
       }
     ],
@@ -208,5 +208,5 @@
       }
     ]
   },
-  "generated_at": "2023-01-10T15:49:47Z"
+  "generated_at": "2023-01-10T18:48:07Z"
 }
